+++ src/components/CreateCourse/components/AuthorItem/AuthorItem.tsx
@@ -1,6 +1,13 @@
 import React, { FC } from 'react';
 import Button from 'src/components/common/Button/Button';
-import { AuthorItemProps } from 'src/types/common/types';
+import { Author } from 'src/types/common/types';
+
+export interface AuthorItemProps {
+	author: Author;
+	isInChosenList: boolean;
+	handleDeletion?: (author: Author) => void;
+	addAuthor?: (author: Author) => void;
+}
 
 const AuthorItem: FC<AuthorItemProps> = ({
 	author,
