+++ src/types/common/types.ts
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export interface Author {
 	id: string;
 	name: string;
@@ -14,25 +12,10 @@ export interface Course {
 	authors: string[];
 }
 
-export interface CourseCardProps {
-	course: Course;
-}
-export interface AuthorItemProps {
-	author: Author;
-	isInChosenList: boolean;
-	handleDeletion?: (author: Author) => void;
-	addAuthor?: (author: Author) => void;
-}
+export type CourseCreated = Omit<Course, 'id'>;
 
-export interface ButtonProps extends ButtonAttributes {
-	buttonText: string;
-	className?: string;
-	style?: React.CSSProperties;
-	handleClick?: (() => void) | ((event: React.FormEvent) => Promise<void>);
-}
-
-// Shared interface for common properties
-interface CommonProps {
+// Shared interface for common Input properties
+export interface InputCommonProps {
 	/**
 	 * The error object for validations
 	 */
@@ -42,46 +25,11 @@ interface CommonProps {
 	 */
 	labelName?: string;
 }
-export type InputProps = Omit<
-	React.InputHTMLAttributes<HTMLInputElement>,
-	'prefix'
-> &
-	CommonProps;
 
-// Textarea component props
-export type TextareaProps = Omit<
-	React.TextareaHTMLAttributes<HTMLTextAreaElement>,
-	'prefix'
-> &
-	CommonProps;
-
-export type ButtonAttributes = Omit<
-	React.ButtonHTMLAttributes<HTMLButtonElement>,
-	'prefix'
->;
-export interface FooterFormProps {
-	duration: string;
-	errors: { [key: string]: string };
-	inputChange: (
-		event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-	) => void;
-}
-
-export interface HeaderFormProps {
-	title: string;
-	inputChange: (
-		event: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-	) => void;
-	description: string;
-	errors: { [key: string]: string };
-}
-
-export interface HeaderProps {
-	userName: string;
-}
-
-export interface User {
+export interface AccountInfo {
 	name: string;
 	email: string;
 	password: string;
 }
+
+export type User = Omit<AccountInfo, 'password'>;
